Move source files from utility to nonce
