+++ include/uit/conduit/proc/inlet/accumulating+type=fundamental/f::AccumulateDuct.hpp
@@ -14,7 +14,7 @@
 #include "../../../../distributed/RdmaWindowManager.hpp"
 #include "../../../../distributed/RdmaAccumulatorPacket.hpp"
 #include "../../../../mpi/Request.hpp"
-#include "../../../../utility/CircularIndex.hpp"
+#include "../../../../nonce/CircularIndex.hpp"
 #include "../../../../utility/identity.hpp"
 #include "../../../../utility/print_utils.hpp"
 #include "../../../../debug/WarnOnce.hpp"
