+++ include/uit/conduit/proc/inlet/put=dropping+type=span/impl/SpanRingImmediateSendDuct.hpp
@@ -15,7 +15,7 @@
 #include "../../../../../debug/err_audit.hpp"
 #include "../../../../../mpi/mpi_utils.hpp"
 #include "../../../../../mpi/Request.hpp"
-#include "../../../../../utility/CircularIndex.hpp"
+#include "../../../../../nonce/CircularIndex.hpp"
 #include "../../../../../utility/print_utils.hpp"
 #include "../../../../../datastructs/RingBuffer.hpp"
