+++ include/uit/conduit/proc/inlet/put=growing+type=cereal/impl/CerealBlockingSendDuct.hpp
@@ -13,7 +13,7 @@
 #include "../../../../../../../third-party/Empirical/source/tools/string_utils.h"
 
 #include "../../../../../mpi/mpi_utils.hpp"
-#include "../../../../../utility/CircularIndex.hpp"
+#include "../../../../../nonce/CircularIndex.hpp"
 #include "../../../../../utility/identity.hpp"
 #include "../../../../../utility/print_utils.hpp"
