+++ include/uit/conduit/proc/inlet/put=growing+type=cereal/impl/CerealDequeImmediateSendDuct.hpp
@@ -15,7 +15,7 @@
 
 #include "../../../../../mpi/mpi_utils.hpp"
 #include "../../../../../mpi/Request.hpp"
-#include "../../../../../utility/CircularIndex.hpp"
+#include "../../../../../nonce/CircularIndex.hpp"
 #include "../../../../../utility/identity.hpp"
 #include "../../../../../utility/print_utils.hpp"
