+++ include/uit/conduit/proc/inlet/put=growing+type=trivial/impl/TrivialDequeImmediateSendDuct.hpp
@@ -13,7 +13,7 @@
 
 #include "../../../../../mpi/mpi_utils.hpp"
 #include "../../../../../mpi/Request.hpp"
-#include "../../../../../utility/CircularIndex.hpp"
+#include "../../../../../nonce/CircularIndex.hpp"
 #include "../../../../../utility/identity.hpp"
 #include "../../../../../utility/print_utils.hpp"
