+++ include/uit/conduit/proc/outlet/get=skipping+type=span/s::BlockIrecvDuct.hpp
@@ -12,7 +12,7 @@
 #include "../../../../debug/WarnOnce.hpp"
 #include "../../../../mpi/mpi_utils.hpp"
 #include "../../../../mpi/Request.hpp"
-#include "../../../../utility/CircularIndex.hpp"
+#include "../../../../nonce/CircularIndex.hpp"
 #include "../../../../utility/identity.hpp"
 #include "../../../../utility/print_utils.hpp"
