+++ include/uit/conduit/proc/outlet/get=skipping+type=trivial/t::BlockIrecvDuct.hpp
@@ -11,7 +11,7 @@
 
 #include "../../../../mpi/mpi_utils.hpp"
 #include "../../../../mpi/Request.hpp"
-#include "../../../../utility/CircularIndex.hpp"
+#include "../../../../nonce/CircularIndex.hpp"
 #include "../../../../utility/identity.hpp"
 #include "../../../../utility/print_utils.hpp"
