+++ include/uit/conduit/proc/outlet/get=stepping+type=span/s::IprobeDuct.hpp
@@ -15,7 +15,7 @@
 
 #include "../../../../mpi/mpi_utils.hpp"
 #include "../../../../mpi/Request.hpp"
-#include "../../../../utility/CircularIndex.hpp"
+#include "../../../../nonce/CircularIndex.hpp"
 #include "../../../../utility/identity.hpp"
 #include "../../../../utility/imemstream.hpp"
 #include "../../../../utility/print_utils.hpp"
