+++ macrobenchmarks/conduit/intra/IntraDuct.hpp
@@ -12,7 +12,7 @@
 #include "uit/polyfill/latch.hpp"
 #include "uit/topology/RingTopologyFactory.hpp"
 #include "uit/debug/benchmark_utils.hpp"
-#include "uit/utility/CircularIndex.hpp"
+#include "uit/nonce/CircularIndex.hpp"
 #include "uit/debug/safe_cast.hpp"
 #include "uit/utility/TimeGuard.hpp"
