+++ macrobenchmarks/conduit/thread/ThreadDuct.hpp
@@ -13,7 +13,7 @@
 #include "uit/polyfill/latch.hpp"
 #include "uit/topology/RingTopologyFactory.hpp"
 #include "uit/debug/benchmark_utils.hpp"
-#include "uit/utility/CircularIndex.hpp"
+#include "uit/nonce/CircularIndex.hpp"
 #include "uit/debug/safe_cast.hpp"
 #include "uit/utility/TimeGuard.hpp"
