+++ microbenchmarks/conduit/DuctBenchmarkRunner.hpp
@@ -17,7 +17,7 @@
 #include "uit/parallel/ThreadTeam.hpp"
 #include "uit/parallel/thread_utils.hpp"
 #include "uit/debug/benchmark_utils.hpp"
-#include "uit/utility/CircularIndex.hpp"
+#include "uit/nonce/CircularIndex.hpp"
 #include "uit/debug/safe_cast.hpp"
 #include "uit/utility/TimeGuard.hpp"
