+++ microbenchmarks/conduit/proc/ProcDuct.hpp
@@ -1,6 +1,6 @@
 #include "uit/meta/ForEach.hpp"
 #include "uit/mpi/MpiGuard.hpp"
-#include "uit/utility/ScopeGuard.hpp"
+#include "uit/nonce/ScopeGuard.hpp"
 
 #include "../DuctBenchmarkRegistration.hpp"
