+++ microbenchmarks/mpi/MPI_Irecv.cpp
@@ -7,7 +7,7 @@
 #include "uit/mpi/MpiGuard.hpp"
 #include "uit/mpi/mpi_utils.hpp"
 #include "uit/debug/benchmark_utils.hpp"
-#include "uit/utility/ScopeGuard.hpp"
+#include "uit/nonce/ScopeGuard.hpp"
 
 const uit::MpiGuard guard;
