+++ tests/Makefile
@@ -13,6 +13,7 @@ TARGET_NAMES += math
 TARGET_NAMES += mesh
 TARGET_NAMES += meta
 TARGET_NAMES += mpi
+TARGET_NAMES += nonce
 TARGET_NAMES += topology
 TARGET_NAMES += utility
