+++ tests/conduit/proc/ProcDuct.hpp
@@ -23,7 +23,7 @@
 #include "uit/topology/ProConTopologyFactory.hpp"
 #include "uit/topology/RingTopologyFactory.hpp"
 #include "uit/utility/assign_utils.hpp"
-#include "uit/utility/CircularIndex.hpp"
+#include "uit/nonce/CircularIndex.hpp"
 #include "uit/math/math_utils.hpp"
 #include "uit/debug/safe_cast.hpp"
