+++ tests/nonce/Makefile
@@ -1,9 +1,8 @@
 SHELL := /bin/bash
 
-TARGET_NAMES += imemstream
-TARGET_NAMES += NamedArrayElement
-TARGET_NAMES += Uninitialized
-TARGET_NAMES += ValueInitialized
+TARGET_NAMES += CircularIndex
+TARGET_NAMES += ScopeGuard
+TARGET_NAMES += spector
 
 TO_ROOT := $(shell git rev-parse --show-cdup)
